+++ RELEASE_NOTES.md
@@ -44,4 +44,4 @@
 #### 2.0.172 - August, 20th
 * Update latest Terminal.Gui (1.72)
 * Bugfix Messagebox-Helper Exception on Cancel Dialog
-* Added the Wizard (Example under Dialogs)
+* Added the Wizard (Example under Wizard)